requirements.txt
flask
uWSGI
flask-argon2
Flask-JWT
PyMySQL
mysql-connector-python
sqlalchemy